Moved blitImageRect() from Atlas to Image.
