+++ src/image.hpp
@@ -8,7 +8,7 @@ class Image
 {
 public:
     explicit Image(const std::filesystem::path& filepath, bool deferLoading = false);
-    explicit Image(size_t rows, size_t cols, size_t channels);
+    explicit Image(size_t width, size_t height, size_t channels);
 
     void allocate(size_t width, size_t height, size_t channels);
     void release();
@@ -18,11 +18,13 @@ public:
     void load();
     void save(const std::filesystem::path& filename) const;
 
-    uint8_t* data(size_t row, size_t col, size_t channel) const;
+    uint8_t* data(size_t x, size_t y, size_t channel) const;
 
     size_t channels() const;
     size_t height() const;
     size_t width() const;
+
+    void blitImage(const Image& source, size_t x, size_t y);
 
 private:
     void load(const std::filesystem::path& filepath);
