+++ src/image_stb.cpp
@@ -1,36 +1,10 @@
 #include <cstring>
 #include <filesystem>
 #include <stdexcept>
-#include <string>
 
 #include "image.hpp"
 #include "vendor/stb/stb_image.h"
 #include "vendor/stb/stb_image_write.h"
-
-Image::Image(const std::filesystem::path& filepath, bool deferLoading)
-{
-    if (deferLoading)
-    {
-        loadinfo(filepath);
-    }
-    else
-    {
-        load(filepath);
-    }
-}
-
-Image::Image(size_t width, size_t height, size_t channels)
-{
-    allocate(width, height, channels);
-}
-
-void Image::allocate(size_t width, size_t height, size_t channels)
-{
-    mWidth = width;
-    mHeight = height;
-    mChannels = channels;
-    mData = std::make_shared<std::vector<uint8_t>>(mWidth * mHeight * mChannels);
-}
 
 void Image::loadinfo(const std::filesystem::path& filepath)
 {
@@ -75,46 +49,7 @@ void Image::load(const std::filesystem::path& filepath)
     stbi_image_free(buffer);
 }
 
-void Image::load()
-{
-    load(mFilePath);
-}
-
-uint8_t* Image::data(size_t row, size_t col, size_t channel) const
-{
-    if (!mData)
-    {
-        return nullptr;
-    }
-
-    size_t offset = row * mWidth * mChannels + col * mChannels + channel;
-    return (*mData).data() + offset;
-}
-
-size_t Image::channels() const
-{
-    return mChannels;
-}
-
-size_t Image::height() const
-{
-    return mHeight;
-}
-
-size_t Image::width() const
-{
-    return mWidth;
-}
-
 void Image::save(const std::filesystem::path& filepath) const
 {
     stbi_write_png(filepath.string().c_str(), mWidth, mHeight, mChannels, (*mData).data(), mWidth * mChannels);
-}
-
-void Image::release()
-{
-    mData = nullptr;
-    mChannels = 0;
-    mHeight = 0;
-    mWidth = 0;
 }
